refactor: vertically organize PeerSharing
